fix import and reorganize more
