+++ e2e/test/scenarios/admin/datamodel/datamodel.cy.spec.js
@@ -2,9 +2,9 @@ import {
   SAMPLE_DB_ID,
   SAMPLE_DB_SCHEMA_ID,
   WRITABLE_DB_ID,
-  ORDERS_QUESTION_ID,
 } from "e2e/support/cypress_data";
 import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
+import { ORDERS_QUESTION_ID } from "e2e/support/cypress_sample_instance_data";
 import {
   restore,
   visitAlias,
